requirements_no_uvloop.txt
# =============================================================================
# DATA QUERYING AI - ESSENTIAL REQUIREMENTS (NO UVLOOP)
# =============================================================================
# Core dependencies for the Data Querying AI project without uvloop
# (uvloop is not supported on Windows)

# =============================================================================
# CORE DATA PROCESSING
# =============================================================================
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0

# =============================================================================
# LANGCHAIN ECOSYSTEM
# =============================================================================
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=2.1.5
langgraph>=0.1.0

# =============================================================================
# AI/ML APIs
# =============================================================================
openai>=1.0.0
google-generativeai>=0.8.5

# =============================================================================
# VECTOR DATABASE
# =============================================================================
qdrant-client>=1.6.0

# =============================================================================
# CONFIGURATION & ENVIRONMENT
# =============================================================================
python-dotenv>=0.19.0
pyyaml>=6.0
pydantic>=2.0.0

# =============================================================================
# HTTP REQUESTS
# =============================================================================
requests>=2.28.0

# =============================================================================
# DATA VISUALIZATION
# =============================================================================
matplotlib>=3.5.0
seaborn>=0.11.0

# =============================================================================
# FILE PROCESSING
# =============================================================================
openpyxl>=3.0.0

# =============================================================================
# TESTING
# =============================================================================
pytest>=7.0.0

# =============================================================================
# INTUGLE DEPENDENCIES
# =============================================================================
# Essential dependencies for Intugle integration
# Note: Intugle should be installed from local source: cd data-tools && pip install -e .

# Core Intugle dependencies
nltk>=3.9.1
asyncpg>=0.30.0
fastapi[standard]>=0.116.1
mcp[cli]>=1.12.4
networkx>=3.4.2
pyaml>=25.7.0
pydantic-settings>=2.10.1
pyfunctional>=1.5.0
symspellpy>=6.9.0
trieregex>=1.0.0
xgboost<=3.0.4
duckdb>=1.3.2
rich>=14.1.0

# =============================================================================
# INSTRUCTIONS
# =============================================================================
# Intugle (install from local source):
# Run: cd data-tools && pip install -e .

# Note: uvloop is not included as it's not supported on Windows
